backend/requirements.txt
langchain
openai

langchain_openai

langchain_community
langchain_chroma
unstructured
markdown
faiss-cpu
langchainhub
psycopg2-binary
redis
cryptography
fastapi
python-multipart